Refactor LocalOnlyHotspotService
