+++ mobile/src/main/java/be/mygod/vpnhotspot/net/TetheringManager.kt
@@ -125,11 +125,8 @@ object TetheringManager {
         stopTethering.invoke(app.connectivity, type)
     }
 
-    fun getAllTetheredIfaces(extras: Bundle) = if (Build.VERSION.SDK_INT >= 26)
-        extras.getStringArrayList(EXTRA_ACTIVE_TETHER).toSet() + extras.getStringArrayList(EXTRA_ACTIVE_LOCAL_ONLY)
-    else extras.getStringArrayList(EXTRA_ACTIVE_TETHER_LEGACY).toSet()
     fun getTetheredIfaces(extras: Bundle) = extras.getStringArrayList(
             if (Build.VERSION.SDK_INT >= 26) EXTRA_ACTIVE_TETHER else EXTRA_ACTIVE_TETHER_LEGACY)
-    @RequiresApi(26)
-    fun getLocalOnlyTetheredIfaces(extras: Bundle) = extras.getStringArrayList(EXTRA_ACTIVE_LOCAL_ONLY)
+    fun getLocalOnlyTetheredIfaces(extras: Bundle) =
+            if (Build.VERSION.SDK_INT >= 26) extras.getStringArrayList(EXTRA_ACTIVE_LOCAL_ONLY) else emptyList<String>()
 }
